python/pyproject.toml
[project]
name = "semantic-kernel-demos"
version = "1.0.0"
description = "Python demos for Semantic Kernel"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "semantic-kernel>=1.35.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
sk-demos = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
